+++ src/main/java/com/trblunt/DrPepper/types/Vitals.java
@@ -2,13 +2,13 @@ package com.trblunt.DrPepper.types;
 
 public class Vitals {
     public int height;
-    public int weight;
-    public int temp;
+    public double weight;
+    public double temp;
     public String bloodPressure;
     // private boolean overTwelve;  // don't know where we want to implement this 
     public String allergies;
 
-    public Vitals(int height, int weight, int temp, String bloodPressure, String allergies) {
+    public Vitals(int height, double weight, double temp, String bloodPressure, String allergies) {
         this.height = height;
         this.weight = weight;
         this.temp = temp;
